improved code by applying ideas from another PR
